# Use atomics instead of synchronized for Gauge