--- a/simpleclient/src/main/java/io/prometheus/client/Gauge.java
+++ b/simpleclient/src/main/java/io/prometheus/client/Gauge.java
@@ -135,7 +135,7 @@ public class Gauge extends SimpleCollector<Gauge.Child> implements Collector.Des
    */
   public static class Child {
 
-    private volatile DoubleAdder value = new DoubleAdder();
+    private final DoubleAdder value = new DoubleAdder();
 
     static TimeProvider timeProvider = new TimeProvider();
 
@@ -167,13 +167,9 @@ public class Gauge extends SimpleCollector<Gauge.Child> implements Collector.Des
      * Set the gauge to the given value.
      */
     public void set(double val) {
-      DoubleAdder updated = new DoubleAdder();
-      updated.add(val);
-
-      // On concurrent `get` and `set`, it is acceptable to `get` an outdated `value`.
       // On concurrent `set` and `set`, it should be acceptable to lose one `set` measurement.
       // On concurrent `set` and `inc / dec`, it might be acceptable to lose the `inc / dec` measurement.
-      value = updated;
+      value.set(val);
     }
     /**
      * Set the gauge to the current unixtime.
@@ -235,6 +231,9 @@ public class Gauge extends SimpleCollector<Gauge.Child> implements Collector.Des
      * Get the value of the gauge.
      */
     public double get() {
+      // On concurrent `get` and `set`, it is acceptable to `get` an outdated `value`.
+      // On concurrent `get` and `inc / dec`, it is acceptable to `get` an outdated `value`.
+      // On concurrent `get` and `set` and `inc / dec`, it is possible to `get` an invalid `value`.
       return value.sum();
     }
   }