--- a/simpleclient/src/main/java/io/prometheus/client/Gauge.java
+++ b/simpleclient/src/main/java/io/prometheus/client/Gauge.java
@@ -134,7 +134,8 @@ public class Gauge extends SimpleCollector<Gauge.Child> implements Collector.Des
    * {@link SimpleCollector#remove} or {@link SimpleCollector#clear},
    */
   public static class Child {
-    private final DoubleAdder value = new DoubleAdder();
+
+    private volatile DoubleAdder value = new DoubleAdder();
 
     static TimeProvider timeProvider = new TimeProvider();
 
@@ -166,13 +167,13 @@ public class Gauge extends SimpleCollector<Gauge.Child> implements Collector.Des
      * Set the gauge to the given value.
      */
     public void set(double val) {
-      synchronized(this) {
-        value.reset();
-        // If get() were called here it'd see an invalid value, so use a lock.
-        // inc()/dec() don't need locks, as all the possible outcomes
-        // are still possible if set() were atomic so no new races are introduced.
-        value.add(val);
-      }
+      DoubleAdder updated = new DoubleAdder();
+      updated.add(val);
+
+      // On concurrent `get` and `set`, it is acceptable to `get` an outdated `value`.
+      // On concurrent `set` and `set`, it should be acceptable to lose one `set` measurement.
+      // On concurrent `set` and `inc / dec`, it might be acceptable to lose the `inc / dec` measurement.
+      value = updated;
     }
     /**
      * Set the gauge to the current unixtime.
@@ -234,9 +235,7 @@ public class Gauge extends SimpleCollector<Gauge.Child> implements Collector.Des
      * Get the value of the gauge.
      */
     public double get() {
-      synchronized(this) {
-        return value.sum();
-      }
+      return value.sum();
     }
   }