Idea:
> We could implement `Gauge#set` with initializing a `new DoubleAdder` to avoid exposing invalid state, but that might have significant allocation (GC) overhead.

Pros:
* avoids thread contention for highly concurrent workloads (`get`, `set`)
* reduces constant overhead (`get`, `set`)

Cons:
* introduces allocations (`set` and `inc / dec`)
* increases GC overhead for high throughput workloads (`set` and `inc / dec`)

Relates to #480