@rrakos-evo 2019-06-05
I ran multiple benchmarks with `24 threads` and picked the result with the smallest error, but of course I can’t know how it behaves under different workloads.

Benchmark:
```
# JMH version: 1.21
# VM version: JDK 11.0.2, OpenJDK 64-Bit Server VM, 11.0.2+9
# VM invoker: /Library/Java/JavaVirtualMachines/openjdk-11.0.2.jdk/Contents/Home/bin/java
# VM options: -javaagent:/Applications/IntelliJ IDEA CE.app/Contents/lib/idea_rt.jar=49590:/Applications/IntelliJ IDEA CE.app/Contents/bin -Dfile.encoding=UTF-8
# Warmup: 5 iterations, 10 s each
# Measurement: 4 iterations, 10 s each
# Timeout: 10 min per iteration
# Threads: 24 threads, will synchronize iterations
```

Before:
```
Benchmark                                                 Mode  Cnt     Score     Error  Units
GaugeBenchmark.prometheusSimpleGaugeChildDecBenchmark     avgt    4    31.720 ±   1.850  ns/op
GaugeBenchmark.prometheusSimpleGaugeChildIncBenchmark     avgt    4    33.041 ±   0.608  ns/op
GaugeBenchmark.prometheusSimpleGaugeChildSetBenchmark     avgt    4  1441.987 ±  29.700  ns/op
GaugeBenchmark.prometheusSimpleGaugeDecBenchmark          avgt    4   154.518 ±  17.714  ns/op
GaugeBenchmark.prometheusSimpleGaugeIncBenchmark          avgt    4   155.453 ±   7.384  ns/op
GaugeBenchmark.prometheusSimpleGaugeSetBenchmark          avgt    4  3047.817 ± 168.989  ns/op
```

After:
```
Benchmark                                                 Mode  Cnt    Score    Error  Units
GaugeBenchmark.prometheusSimpleGaugeChildDecBenchmark     avgt    4   30.618 ±  0.867  ns/op
GaugeBenchmark.prometheusSimpleGaugeChildIncBenchmark     avgt    4   31.543 ±  5.164  ns/op
GaugeBenchmark.prometheusSimpleGaugeChildSetBenchmark     avgt    4  314.646 ±  4.892  ns/op
GaugeBenchmark.prometheusSimpleGaugeDecBenchmark          avgt    4  150.270 ±  9.382  ns/op
GaugeBenchmark.prometheusSimpleGaugeIncBenchmark          avgt    4  156.412 ±  9.367  ns/op
GaugeBenchmark.prometheusSimpleGaugeSetBenchmark          avgt    4  537.230 ± 67.680  ns/op
```