@brian-brazil 2019-06-05
What I was thinking about back when adding this is that we could adjust DoubleAdder to handle this more efficiently if it ever became an issue.