@rrakos-evo 2019-06-05
I’m just thinking out loudly:
We could introduce `DoubleAdder#set` as an atomic alternative to `reset` and `add`.
I'm not sure how to update `base` and `cells` atomically without locking using `busy`.