@rrakos-evo 2019-06-07
`81283a20`:
```
Benchmark                                                 Mode  Cnt    Score    Error  Units
GaugeBenchmark.codahaleCounterDecBenchmark                avgt    4   35.166 ± 16.052  ns/op
GaugeBenchmark.codahaleCounterIncBenchmark                avgt    4   31.423 ± 10.979  ns/op
GaugeBenchmark.prometheusSimpleGaugeChildDecBenchmark     avgt    4   38.966 ±  3.249  ns/op
GaugeBenchmark.prometheusSimpleGaugeChildIncBenchmark     avgt    4   37.798 ±  9.753  ns/op
GaugeBenchmark.prometheusSimpleGaugeChildSetBenchmark     avgt    4  429.808 ± 92.764  ns/op
GaugeBenchmark.prometheusSimpleGaugeDecBenchmark          avgt    4  182.829 ± 38.544  ns/op
GaugeBenchmark.prometheusSimpleGaugeIncBenchmark          avgt    4  184.254 ± 43.754  ns/op
GaugeBenchmark.prometheusSimpleGaugeSetBenchmark          avgt    4  582.121 ± 86.330  ns/op
GaugeBenchmark.prometheusSimpleGaugeNoLabelsDecBenchmark  avgt    4   37.136 ±  6.599  ns/op
GaugeBenchmark.prometheusSimpleGaugeNoLabelsIncBenchmark  avgt    4   39.167 ±  3.439  ns/op
GaugeBenchmark.prometheusSimpleGaugeNoLabelsSetBenchmark  avgt    4  456.339 ± 61.951  ns/op
```