@rrakos-evo 2019-06-13
2874c5ee:
```
Benchmark                                                 Mode  Cnt    Score    Error  Units
GaugeBenchmark.codahaleCounterDecBenchmark                avgt    4   35.740 ± 20.478  ns/op
GaugeBenchmark.codahaleCounterIncBenchmark                avgt    4   32.936 ±  1.891  ns/op
GaugeBenchmark.prometheusSimpleGaugeChildDecBenchmark     avgt    4   36.128 ±  3.627  ns/op
GaugeBenchmark.prometheusSimpleGaugeChildIncBenchmark     avgt    4   38.462 ±  1.829  ns/op
GaugeBenchmark.prometheusSimpleGaugeChildSetBenchmark     avgt    4  444.700 ± 97.494  ns/op
GaugeBenchmark.prometheusSimpleGaugeDecBenchmark          avgt    4  163.684 ±  4.272  ns/op
GaugeBenchmark.prometheusSimpleGaugeIncBenchmark          avgt    4  161.102 ± 12.008  ns/op
GaugeBenchmark.prometheusSimpleGaugeSetBenchmark          avgt    4  492.110 ± 50.022  ns/op
GaugeBenchmark.prometheusSimpleGaugeNoLabelsDecBenchmark  avgt    4   37.190 ± 12.319  ns/op
GaugeBenchmark.prometheusSimpleGaugeNoLabelsIncBenchmark  avgt    4   35.756 ±  5.226  ns/op
GaugeBenchmark.prometheusSimpleGaugeNoLabelsSetBenchmark  avgt    4  471.402 ± 92.968  ns/op
```