@rrakos-evo 2019-08-15
I couldn't find `Gauge` in YourKit snapshots and screenshots from July. 🙁 
I'll need to profile the changes again, so I'll try to prove that it eliminated blocking.