@rrakos-evo 2019-09-05
I can confirm that there's no more blocking. I've been trying to get profiling results to have something to show that proves that, but because it's so fast now it doesn't show up during sampling using a reasonable sampling rate, that does not affect performance by itself.

Do you have any further concerns about merging this?